+++ config_joker/utils/parser.py
@@ -1,4 +1,4 @@
-from typing import Union, List, Any
+from typing import Union, List, Any, Dict
 import regex as re
 
 
@@ -34,3 +34,20 @@ def dict_extractor(path: str, data: str) -> Any:
     for key in parsed_path:
         data = data[key]
     return data
+
+
+class MissingKeyNameInListOfDicts(Exception):
+    def __init__(self, k: str, v: str, data: str) -> None:
+        msg = f'Components missing keys "{k}" or "{v}" in data: {data}'
+        super().__init__(msg)
+
+
+def parse_list_of_dicts(data: List[Dict], k_name: str, v_name: str) -> Dict:
+    result = {}
+    if not isinstance(data, list):
+        raise TypeError(f'data must be a list, value passed: {data}')
+    try:
+        _ = [result.update({d[k_name]: d[v_name]}) for d in data]
+    except KeyError:
+        raise MissingKeyNameInListOfDicts(k=k_name, v=v_name, data=data)
+    return result